api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.75.0
